Add jump table implementation
Since Rust does not support the goto label mechanism similar to C
language, it cannot directly implement the jump table in Linux. We
implement it by defining a naked function and modifying the instructions
in the function.

This method does not reduce the number of instructions and branch
judgments, but it can reduce the number of memory accesses.