Add tests for global variables of eBPF programs
The processing of global variables in eBPF programs is supported by Map.
The user library needs to process the eBPF executable file to create the
corresponding map and relocate the instructions.

With reference to Aya, we implemented a simple loader and executor. The
loader is used in user mode, and its function is to parse the eBPF
program, call the system call to create and manage the map, and modify
the instruction according to the relocation information. The current
call system call is a virtual function.

The executor is used in kernel mode, and its function is to modify some
instructions that load data from the map again.

Due to limitations in the rbpf interpreter implementation, we
temporarily turned off some checks.